Add warning to empty-command and add warn command

+++ src/script/packetdecoder/command/EmptyCommand.java
@@ -1,6 +1,9 @@
 package script.packetdecoder.command;
 
 import com.palidino.osrs.io.Command;
+import com.palidino.osrs.model.dialogue.Dialogue;
+import com.palidino.osrs.model.dialogue.DialogueEntry;
+import com.palidino.osrs.model.dialogue.DialogueScript;
 import com.palidino.osrs.model.item.Item;
 import com.palidino.osrs.model.player.Player;
 import lombok.var;
@@ -14,10 +17,19 @@ public class EmptyCommand implements Command {
 
     @Override
     public void execute(Player player, String message) {
-        for (var i = 0; i < player.getInventory().size(); i++) {
-            var id = player.getInventory().getId(i);
-            player.getInventory().deleteItem(id, Item.MAX_AMOUNT);
-        }
-        player.getGameEncoder().sendMessage("You empty your inventory..");
+        DialogueEntry entry = new DialogueEntry();
+        entry.setSelection("Are you sure you want to empty your inventory?", "Yes, empty my inventory!", "No!");
+        DialogueScript script = (p, index, childId, slot) -> {
+            for (var i = 0; i < player.getInventory().size(); i++) {
+                var id = player.getInventory().getId(i);
+                if (slot == 0) {
+                    player.getInventory().deleteItem(id, Item.MAX_AMOUNT);
+                    player.getGameEncoder().sendMessage("You empty your inventory..");
+                } else {
+                    return;
+                }
+            }
+        };
+        Dialogue.open(player, entry, script);
     }
 }
